Increase size of pthreads types on 32-bit systems.
Remove dependence on libatomic stubs by using __sync* primitives
instead of __atomic ones on 32-bit systems. 64-bit systems should be
the same as before.

Fix bug where work_queue init would overwrite the mutex locking the
work_queue.

+++ src/runtime/posix_threads.cpp
@@ -6,17 +6,15 @@
 
 extern "C" {
 
-// On posix platforms, there's a 1-to-1 correspondence between
-// halide_* threading functions and the pthread_* functions. We take
-// some liberties with the types of the opaque pointer objects to
-// avoid a bunch of pointer casts.
-
+// This code cannot depend on system headers, hence we choose a data size which will
+// be large enough for all systems we care about.
+// 64 bytes covers this for both mutex and condvar. Using int64_t ensures alignment.
 struct pthread_mutex_t {
-    uintptr_t _private[8];
+    int64_t _private[8];
 };
 
 struct pthread_cond_t {
-    uintptr_t _private[8];
+    int64_t _private[8];
 };
 
 typedef long pthread_t;
